requirements.txt
fastapi
uvicorn[standard]
google-generativeai
google-api-python-client
youtube-transcript-api
python-dotenv
pydantic
